+++ src/views/Course.vue
@@ -21,11 +21,19 @@
         </div>
       </div>
       <div class="column is-4">
-        <div class="box">
-          <h4 class="title is-4">Links</h4>
+        <div class="box has-text-left">
           <aside class="menu">
+            <p class="menu-label">
+              Main Links
+            </p>
             <ul class="menu-list">
               <li><a :href="'https://' + server.domain + '/courses/' + course.id">Canvas Link</a></li>
+              <li v-for="(name, link) in autoLinks" :key="link"><a :href="link">{{name}}</a></li>
+            </ul>
+            <p class="menu-label">
+              User Created Links
+            </p>
+            <ul class="menu-list">
               <li v-for="(name, link) in links" :key="link"><a :href="link">{{name}}</a></li>
             </ul>
           </aside>
@@ -45,6 +53,7 @@
           <h5 class="title is-5">Utilities</h5>
           <div class="buttons">
             <b-button type="is-info" :loading="zoomFinderLoading" icon-left="video" @click="findZoomLink">Find Zoom Link</b-button>
+            <b-button style="background-color: #7289da; color: white" :loading="discordFinderLoading" icon-left="discord" @click="findDiscordLink">Find Discord Link</b-button>
           </div>
         </div>
       </div>
@@ -59,18 +68,22 @@ export default {
     const courseInfo = window.localStorage.canvas_course_info ? JSON.parse(window.localStorage.canvas_course_info) : {}
     if(courseInfo[this.course.id]) {
       this.links = courseInfo[this.course.id].links
+      this.autoLinks = courseInfo[this.course.id].autoLinks || {}
+
     }
     this.debounceSave = debounce(this, this.save, 300)
   },
   data() {
     return {
-      links: {}, //Map of <URL, Name>
+      links: {}, //Map of <URL, Name>,
+      autoLinks: {}, //Any links generated automatically (zoom, discord, etc)
       formAddLink: {
         url: null,
         name: null
       },
       notes: null,
       zoomFinderLoading: false,
+      discordFinderLoading: false,
       debounceSave: null
     }
   },
@@ -94,85 +107,128 @@ export default {
     save() {
       console.debug('save')
       let courseInfo = window.localStorage.canvas_course_info ? JSON.parse(window.localStorage.canvas_course_info) : {}
-      if(!courseInfo[this.course.id]) courseInfo[this.course.id] = {links: []}
       courseInfo[this.course.id] = {
         links: this.links,
-        notes: this.notes
+        notes: this.notes,
+        autoLinks: this.autoLinks,
       }
       window.localStorage.canvas_course_info = JSON.stringify(courseInfo)
     },
     async findZoomLink() {
       this.zoomFinderLoading = true;
       try {
-        const link = await searchZoomLink(this.course.id, this.server)
+        const [link, id] = await searchForLink(this.course.id, this.server, ZOOM_LINK_REGEX, 2)
         this.zoomFinderLoading = false
-        if(link) {
+        if(link && id) {
           if(!this.links[link]) {
-            this.$set(this.links, link, "Zoom Link")
+            this.$set(this.autoLinks, link, `Zoom Link (${id})`)
             this.save()
           }
         }else {
           alert('Sorry, no zoom links could be found')
         }
       }catch(err) {
-        alert('Sorry, no zoom links could be found due to an error.')
         console.error('Could not find a zoom link', err)
+        alert('Sorry, no zoom links could be found due to an error.')
         this.zoomFinderLoading = false
+      }
+    },
+    async findDiscordLink() {
+      this.discordFinderLoading = true;
+      try {
+        const [link, id] = await searchForLink(this.course.id, this.server, DISCORD_LINK_REGEX, 1)
+        this.discordFinderLoading = false
+        if(link) {
+          if(!this.links[link]) {
+            this.$set(this.autoLinks, link, `Discord Link (${id})`)
+            this.save()
+          }
+        }else {
+          alert('Sorry, no discord links could be found')
+        }
+      }catch(err) {
+        console.error('Could not find a discord link', err)
+        alert('Sorry, no discord links could be found due to an error.')
+        this.discordFinderLoading = false
       }
     },
   }
 }
 const ZOOM_LINK_REGEX = new RegExp(/https?:\/\/.*?\.?zoom.us\/(j|my)\/([a-z0-9-_]+)/im)
-const PAGE_TITLE_REGEX = new RegExp(/(zoom)|(link)/i)
+const DISCORD_LINK_REGEX = new RegExp(/discord(?:\.com|app\.com|\.gg)\/(?:invite\/)?([a-zA-Z0-9\\-]{2,32})/);
+const PAGE_TITLE_REGEX = new RegExp(/(zoom)|(link)|(discord)/i)
 
-async function searchZoomLink(courseID, server) {
+async function searchForLink(courseID, server, URL_REGEX, regexIDIndex) {
   const options = {headers: {'Authorization': "Bearer " + server.api}}
 
   let response, json;
-
+  console.debug('Scanning for URL', URL_REGEX, regexIDIndex, 'on', server.domain, 'for course', courseID)
+  console.debug('Scanning tabs')
   //Search via TABS
   response = await fetch(server.url + `/api/v1/courses/${courseID}/tabs`, options)
-  json = (await response.json()).find(tab => ZOOM_LINK_REGEX.test(tab))
-  if(json) return json;
+  json = (await response.json()).find(tab => tab.type === "external" && URL_REGEX.test(tab.url))
+  if(json) {
+    const match = json.url.match(URL_REGEX)
+    return [ match[0], match[regexIDIndex] ]
+  }
 
+  console.debug('Scanning pages')
   //Search via PAGES (Needs to view each page.)
   response = await fetch(server.url + `/api/v1/courses/${courseID}/pages`, options)
   if(response.status == 200) {
     json = (await response.json()).filter(page => PAGE_TITLE_REGEX.test(page.title))
     //Find a valid page:
     for(const page of json) {
+      console.debug('Searching page', page.title)
       const response = await fetch(server.url + `/api/v1/courses/${courseID}/pages/${page.url}`, options)
       const json = await response.json()
 
-      const match = json.body.match(ZOOM_LINK_REGEX)
-      if(match && match.length === 3) {
-        return match[0]
+      const match = json.body.match(URL_REGEX)
+      if(match) {
+        return [ match[0], match[regexIDIndex] ]
       }
     }
   }
   
+  console.debug('Scanning modules')
   //Search via MODULES
   response = await fetch(server.url + `/api/v1/courses/${courseID}/modules?include[]=items`, options)
   if(response.status == 200) {
-    json = (await response.json())
+    json = await response.json()
     for(const module of json) {
       for(const item of module.items) {
-        const match = ZOOM_LINK_REGEX.test(item.external_url)
-        if(match) return item.external_url
+        if(item.external_url) {
+          const match = item.external_url.match(URL_REGEX)
+          if(match) return [ item.external_url, match[regexIDIndex] ]
+        }
       }
     }
   }
 
+  console.debug('Scanning syllabus')
   //Find via syllabus
   response = await fetch(server.url + `/api/v1/courses/${courseID}?include[]=syllabus_body`, options)
   if(response.status == 200) {
-    json = (await response.json())
+    json = await response.json()
     
-    const match = json.syllabus_body.match(ZOOM_LINK_REGEX)
-    if(match && match.length === 3) {
-      return match[0]
+    const match = json.syllabus_body.match(URL_REGEX)
+    if(match) {
+      return [ match[0], match[regexIDIndex] ]
     }
   }
+
+  console.debug('Scanning announcements')
+  //find via announcements
+  response = await fetch(server.url + `/api/v1/announcements?context_codes[]=course_${courseID}`, options)
+  if(response.status == 200) {
+    json = await response.json()
+    for(const announcement of json) {
+        const match = announcement.message.match(URL_REGEX)
+        if(match) return [ match[0], match[regexIDIndex] ]
+    }
+  }
+
+  return [null]
 }
 function debounce(context, func, delay) {
   let inDebounce